+++ github2/bin/search_repos.py
@@ -1,6 +1,10 @@
 #! /usr/bin/env python
 # coding: utf-8
 """github_search_repos - search for repositories on GitHub"""
+# Copyright (C) 2011-2012 James Rowe <jnrowe@gmail.com>
+#
+# This file is part of python-github2, and is made available under the 3-clause
+# BSD license.  See LICENSE for the full details.
 
 
 import logging
@@ -17,9 +21,11 @@ PY3K = sys.version_info[0] == 3 and True or False
 
 
 def print_(text):
-    """Python 2 & 3 compatible print function
+    """Python 2 & 3 compatible print function.
 
-    We support <2.6, so can't use __future__.print_function"""
+    We support <2.6, so can't use __future__.print_function
+
+    """
     if PY3K:
         print(text)
     else:
@@ -27,7 +33,7 @@ def print_(text):
 
 
 def parse_commandline():
-    """Parse the comandline and return parsed options."""
+    """Parse the command line and return parsed options."""
 
     parser = OptionParser()
     parser.description = __doc__
@@ -46,7 +52,7 @@ def parse_commandline():
 
 
 def main():
-    """This implements the actual program functionality"""
+    """Implement the actual program functionality."""
     return_value = 0
 
     options, term = parse_commandline()